Refactor tests, run integration tests
